requirements.txt
llama-index>=0.10.0
requests>=2.31.0
pandas>=2.2.0
nltk>=3.8.1
networkx>=3.2.1
pydantic>=2.7.0
pyyaml>=6.0.1
